--- a/priv/repo/migrations/20241114122325_add_index_to_dataclip_id_on_runs.exs
+++ b/priv/repo/migrations/20241114122325_add_index_to_dataclip_id_on_runs.exs
@@ -1,0 +1,8 @@
+defmodule Lightning.Repo.Migrations.AddIndexToDataclipIdOnRuns do
+  use Ecto.Migration
+
+  def change do
+    create_if_not_exists index("runs", [:dataclip_id])
+    create_if_not_exists index("work_orders", [:dataclip_id])
+  end
+end